agent-py/requirements.txt
langchain-openai>=0.2.1
langchain-anthropic>=0.2.1
langchain>=0.3.1
openai>=1.51.0
langchain-community>=0.3.1
copilotkit==0.1.45
uvicorn>=0.31.0
python-dotenv>=1.0.1
langchain-core>=0.3.25
langgraph-cli[inmem]>=0.1.64
langchain-google-genai>=2.1.4,<3.0.0
setuptools>=61.0
